moved customers folder
