+++ packages/kokoas-client/src/pages/projRegisterV2/Contents.tsx
@@ -1,4 +1,4 @@
-import { CustomerSummary } from './sections/CustomerSummary';
+import { CustomerSummary } from './sections/customersSummary/CustomerSummary';
 import { ProjectLocation } from './sections/projectLocation/ProjectLocation';
 import { ProjectInformation } from './sections/projectInformation/ProjectInformation';
 import { PageSubTitle3 } from 'kokoas-client/src/components';
